refactor utils module
so that everything is no longer in __init__.py
