+++ cruft/_commands/create.py
@@ -4,13 +4,8 @@ from typing import Any, Dict, Optional
 
 from cookiecutter.generate import generate_files
 
-from .utils import (
-    example,
-    generate_cookiecutter_context,
-    get_cookiecutter_repo,
-    json_dumps,
-    resolve_template_url,
-)
+from . import utils
+from .utils import example
 
 
 @example("https://github.com/timothycrosley/cookiecutter-python/")
@@ -26,16 +21,18 @@ def create(
     overwrite_if_exists: bool = False,
 ) -> Path:
     """Expand a Git based Cookiecutter template into a new project on disk."""
-    template_git_url = resolve_template_url(template_git_url)
+    template_git_url = utils.cookiecutter.resolve_template_url(template_git_url)
     with TemporaryDirectory() as cookiecutter_template_dir_str:
         cookiecutter_template_dir = Path(cookiecutter_template_dir_str)
-        repo = get_cookiecutter_repo(template_git_url, cookiecutter_template_dir, checkout)
+        repo = utils.cookiecutter.get_cookiecutter_repo(
+            template_git_url, cookiecutter_template_dir, checkout
+        )
         last_commit = repo.head.object.hexsha
 
         if directory:
             cookiecutter_template_dir = cookiecutter_template_dir / directory
 
-        context = generate_cookiecutter_context(
+        context = utils.cookiecutter.generate_cookiecutter_context(
             template_git_url,
             cookiecutter_template_dir,
             config_file,
@@ -56,7 +53,7 @@ def create(
         # After generating the project - save the cruft state
         # into the cruft file.
         (project_dir / ".cruft.json").write_text(
-            json_dumps(
+            utils.cruft.json_dumps(
                 {
                     "template": template_git_url,
                     "commit": last_commit,
