+++ cruft/_commands/link.py
@@ -4,14 +4,8 @@ from typing import Any, Dict, Optional
 
 import typer
 
-from .utils import (
-    example,
-    generate_cookiecutter_context,
-    get_cookiecutter_repo,
-    get_cruft_file,
-    json_dumps,
-    resolve_template_url,
-)
+from . import utils
+from .utils import example
 
 
 @example("https://github.com/timothycrosley/cookiecutter-python/")
@@ -26,17 +20,19 @@ def link(
     directory: Optional[str] = None,
 ) -> bool:
     """Links an existing project created from a template, to the template it was created from."""
-    cruft_file = get_cruft_file(project_dir, exists=False)
-    template_git_url = resolve_template_url(template_git_url)
+    cruft_file = utils.cruft.get_cruft_file(project_dir, exists=False)
+    template_git_url = utils.cookiecutter.resolve_template_url(template_git_url)
     with TemporaryDirectory() as cookiecutter_template_dir_str:
         cookiecutter_template_dir = Path(cookiecutter_template_dir_str)
-        repo = get_cookiecutter_repo(template_git_url, cookiecutter_template_dir, checkout)
+        repo = utils.cookiecutter.get_cookiecutter_repo(
+            template_git_url, cookiecutter_template_dir, checkout
+        )
         last_commit = repo.head.object.hexsha
 
         if directory:
             cookiecutter_template_dir = cookiecutter_template_dir / directory
 
-        context = generate_cookiecutter_context(
+        context = utils.cookiecutter.generate_cookiecutter_context(
             template_git_url,
             cookiecutter_template_dir,
             config_file,
@@ -55,7 +51,7 @@ def link(
             use_commit = typer.prompt("Link to template at commit", default=last_commit)
 
         cruft_file.write_text(
-            json_dumps(
+            utils.cruft.json_dumps(
                 {
                     "template": template_git_url,
                     "commit": use_commit,
