+++ cruft/_commands/utils/diff.py
